Parse your .b6 file into an OTU and taxa table!
For amplicon alignment, the "OTU" table (or tally table of samples vs species) is a well-known output format. This program parses the .b6 into 2 such tables simultaneously (one for which reference was aligned, one optionally for taxonomy). 

It is RAM-friendly and fast. On a 7GB b6 file, it created an OTU and taxa table in 28 seconds using less than 10MB RAM.